backend/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-generativeai
python-jose
passlib


